requirements.txt
# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# API and Web
requests>=2.26.0
beautifulsoup4>=4.9.0
selenium>=4.0.0

# ML and Analysis
scikit-learn>=1.0.0
nltk>=3.6.0
transformers>=4.5.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.1
logging>=0.5.1.2

# Testing
pytest>=6.2.0
